style: switching to biome toolchain
the language tooling has kept changing throughout this project's duration, however this is most definitely the last time it will change. i decided that deno's lsp, no matter the amount of features that it provided, would not be as popular as something like biome, thus causing confusion for future contributors.